+++ src/Utils/SchemaPrinter.php
@@ -338,7 +338,7 @@ class SchemaPrinter
      */
     protected static function printInputValue($arg): string
     {
-        $argDecl = "{$arg->name}: {$arg->getType()->toString()}" . static::printDeprecated($arg);
+        $argDecl = "{$arg->name}: {$arg->getType()->toString()}";
 
         if ($arg->defaultValueExists()) {
             $defaultValueAST = AST::astFromValue($arg->defaultValue, $arg->getType());
@@ -348,10 +348,11 @@ class SchemaPrinter
                 throw new InvariantViolation("Unable to convert defaultValue of argument {$arg->name} into AST: {$inconvertibleDefaultValue}.");
             }
 
-            $argDecl .= ' = ' . Printer::doPrint($defaultValueAST);
+            $printedDefaultValue = Printer::doPrint($defaultValueAST);
+            $argDecl .= " = {$printedDefaultValue}";
         }
 
-        return $argDecl;
+        return $argDecl . static::printDeprecated($arg);
     }
 
     /**