+++ src/Validator/Rules/QueryComplexity.php
@@ -160,6 +160,8 @@ class QueryComplexity extends QuerySecurityRule
     }
 
     /**
+     * Will the given field be executed at all, given the directives placed upon it?
+     *
      * @throws \Exception
      * @throws \ReflectionException
      * @throws InvariantViolation
@@ -219,7 +221,7 @@ class QueryComplexity extends QuerySecurityRule
     }
 
     /** @param array<string, mixed>|null $rawVariableValues */
-    public function setRawVariableValues(array $rawVariableValues = null): void
+    public function setRawVariableValues(?array $rawVariableValues = null): void
     {
         $this->rawVariableValues = $rawVariableValues ?? [];
     }